Deployment/requirements.txt
streamlit
pandas
seaborn
matplotlib
Pillow
numpy
scikit-learn
plotly